Changed the code and renamed the file to encryptUserData form encryptPrivateKey
